Build Process implementation
