+++ .gitignore
@@ -4,3 +4,6 @@ composer.phar
 
 ### IDEs ###
 /.idea/
+
+### Node ###
+/node_modules/
